Codechange: move windows common to AI and GS out of ai_gui
