+++ src/ai/ai_instance.cpp
@@ -14,11 +14,11 @@
 #include "../script/squirrel_class.hpp"
 
 #include "ai_config.hpp"
-#include "ai_gui.hpp"
 #include "ai.hpp"
 
 #include "../script/script_storage.hpp"
 #include "../script/script_cmd.h"
+#include "../script/script_gui.h"
 #include "ai_info.hpp"
 #include "ai_instance.hpp"
 
@@ -64,7 +64,7 @@ void AIInstance::Died()
 	/* Intro is not supposed to use AI, but it may have 'dummy' AI which instant dies. */
 	if (_game_mode == GM_MENU) return;
 
-	ShowAIDebugWindow(_current_company);
+	ShowScriptDebugWindow(_current_company);
 
 	const AIInfo *info = AIConfig::GetConfig(_current_company, AIConfig::SSS_FORCE_GAME)->GetInfo();
 	if (info != nullptr) {
