+++ src/game/game.hpp
@@ -12,10 +12,6 @@
 
 #include "../core/string_compare_type.hpp"
 #include "game_scanner.hpp"
-#include <map>
-
-/** A list that maps AI names to their AIInfo object. */
-typedef std::map<const char *, class ScriptInfo *, StringCompare> ScriptInfoList;
 
 #include "../script/api/script_event_types.hpp"
