+++ src/game/game_instance.cpp
@@ -14,7 +14,7 @@
 
 #include "../script/script_storage.hpp"
 #include "../script/script_cmd.h"
-#include "../ai/ai_gui.hpp"
+#include "../script/script_gui.h"
 #include "game_config.hpp"
 #include "game_info.hpp"
 #include "game_instance.hpp"
@@ -67,7 +67,7 @@ void GameInstance::Died()
 {
 	ScriptInstance::Died();
 
-	ShowAIDebugWindow(OWNER_DEITY);
+	ShowScriptDebugWindow(OWNER_DEITY);
 
 	const GameInfo *info = Game::GetInfo();
 	if (info != nullptr) {
