+++ src/misc.cpp
@@ -11,7 +11,7 @@
 #include "landscape.h"
 #include "news_func.h"
 #include "ai/ai.hpp"
-#include "ai/ai_gui.hpp"
+#include "script/script_gui.h"
 #include "newgrf.h"
 #include "newgrf_house.h"
 #include "economy_func.h"
@@ -100,7 +100,7 @@ void InitializeGame(uint size_x, uint size_y, bool reset_date, bool reset_settin
 	InitializeGraphGui();
 	InitializeObjectGui();
 	InitializeTownGui();
-	InitializeAIGui();
+	InitializeScriptGui();
 	InitializeTrees();
 	InitializeIndustries();
 	InitializeObjects();
