+++ src/table/settings/gui_settings.ini
@@ -779,7 +779,7 @@ cat      = SC_EXPERT
 var      = gui.ai_developer_tools
 flags    = SF_NOT_IN_SAVE | SF_NO_NETWORK_SYNC
 def      = false
-post_cb  = [](auto) { InvalidateWindowClassesData(WC_AI_SETTINGS); }
+post_cb  = [](auto) { InvalidateWindowClassesData(WC_SCRIPT_SETTINGS); }
 cat      = SC_EXPERT
 
 [SDTC_BOOL]
