+++ src/toolbar_gui.h
@@ -41,7 +41,7 @@ enum MainToolbarHotkeys {
 	MTHK_BUILD_TREES,
 	MTHK_MUSIC,
 	MTHK_LANDINFO,
-	MTHK_AI_DEBUG,
+	MTHK_SCRIPT_DEBUG,
 	MTHK_SMALL_SCREENSHOT,
 	MTHK_ZOOMEDIN_SCREENSHOT,
 	MTHK_DEFAULTZOOM_SCREENSHOT,
