+++ src/widgets/game_widget.h
@@ -5,7 +5,7 @@
  * See the GNU General Public License for more details. You should have received a copy of the GNU General Public License along with OpenTTD. If not, see <http://www.gnu.org/licenses/>.
  */
 
-/** @file ai_widget.h Types related to the ai widgets. */
+/** @file game_widget.h Types related to the GS widgets. */
 
 #ifndef WIDGETS_GS_WIDGET_H
 #define WIDGETS_GS_WIDGET_H
